backend/requirements.txt
# Enterprise-Grade HMS Backend Dependencies
# Core Django Framework
Django==4.2.24
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-filter==24.3
django-cors-headers==4.7.0

# Environment and Configuration
python-dotenv==1.1.1

# Database
psycopg2-binary==2.9.10

# Cloud Storage
boto3==1.34.162
django-storages==1.14.4

# Task Queue and Caching
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
django-celery-beat==2.7.0

# Monitoring and Logging
sentry-sdk==2.13.0
whitenoise==6.7.0
django-prometheus==2.3.1

# Security and Cryptography
argon2-cffi==23.1.0
django-cryptography==1.1
cryptography==45.0.7
pyjwt==2.9.0

# Data Processing
openpyxl==3.1.5
lxml==5.3.0
requests==2.32.5
defusedxml==0.7.1

# Circuit Breaker Pattern
pybreaker==1.0.2

# Observability
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-django
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-psycopg2

# Message Queue
kafka-python==2.0.3

# AI/ML and Data Science
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
tensorflow==2.17.0
torch==2.5.1
transformers==4.45.2
spacy==3.7.5
opencv-python==4.10.0.84
Pillow==11.0.0
pytesseract==0.3.13

# Kubernetes and Container Orchestration
kubernetes==29.0.0
docker==7.1.0
python-consul==1.1.0

# Enterprise Monitoring
prometheus-client==0.20.0
grafana-api==1.0.3

# Enterprise Performance
psutil==6.0.0

# Enterprise Communication
grpcio==1.66.1
grpcio-tools==1.66.1

# Enterprise Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
pytest-benchmark==4.0.0
pytest-asyncio==0.25.0
factory-boy==3.3.1
freezegun==1.5.1
responses==0.25.3
httpx==0.28.1

# Performance Testing
locust==2.32.2
psutil==6.0.0
memory-profiler==0.61.0

# Security Testing
bandit==1.8.6
safety==3.6.1
owasp-zap==2.14.0
semgrep==1.99.0

# Accessibility Testing
axe-core==4.10.1
axe-playwright==1.1.0

# Visual Regression Testing
percy==2.0.0
backstopjs==6.3.11

# API Testing
postman-newman==6.2.1
bravado==11.0.3
schemathesis==3.40.5

# Contract Testing
pact-python==2.3.0
pact-foundation==2.2.0

# Cross-Browser Testing
playwright==1.49.0
selenium==4.28.1
browserstack==4.1.0

# E2E Testing
cypress==13.17.0
playwright==1.49.0
testcafe==3.8.0

# Load Testing
k6==0.52.0
artillery==2.0.11
vegeta==12.12.0

# Mobile Testing
appium==4.1.2
maestro==0.5.0
detox==20.28.0

# Database Testing
factory-boy==3.3.1
faker==33.3.1
model-bakery==1.19.0

# Healthcare Compliance
hipaa==0.1.0
gdpr==0.1.0
hl7==0.1.0
fhir==4.2.0

# Code Quality
black==25.1.0
isort==6.0.1
flake8==7.3.0
pylint==3.3.8
mypy==1.18.1
ruff==0.9.1
pre-commit==4.1.0

# Coverage Reporting
coverage==7.6.10
cov-core==1.15.0
cov-report==0.4.0

# Test Data Management
mimesis==17.0.0
 hypothesis==6.123.1
property-based-testing==0.1.0

# Enterprise Utilities
six==1.16.0
python-dateutil==2.9.0.post0
pytz==2024.2

# Encrypted Model Fields
django-encrypted-model-fields==0.6.5

# Code Quality
black==25.1.0
isort==6.0.1
flake8==7.3.0
pylint==3.3.8
mypy==1.18.1
bandit==1.8.6
safety==3.6.1

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.0.1
drf-spectacular-sidecar==2024.12.1
